build(api-server): upgraded Open API generator
Which also lead to an avalanche of other smaller changes that
needed to be done in order to make the whole thing work:
1. The build system is now using different scripts to generate
the typescript code side-stepping the openapi specs coming from
typescript files that get exported to JSON files (instead now
the JSON files are first class citizens under version control
that are meant to be edited directly).
2. Test cases had to be refactored to accomodate for the new
API surface of the generated typescript-axios code.
3. The watch script is now fully operational and does not need
to run a separate process for each package when watching TS code.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>