ci: release automation -> GHCR container image auto-publishing
Refactored the npm publishing workflow to now contain both
the npm and container image (GHCR) publishing jobs.

This was necessary because the easiest way to express the
dependency between the two was via the `needs` syntax
of the container publishing workflow (which must run
after the npm packages have successfully gotten published.

Fixes #27

Signed-off-by: Peter Somogyvari <peter.metz@unarin.com>