fix(cmd-api-server): health-check endpoint crash - no OpenAPI metadata
The issue was that the API server code did a shallow merge on the
OpenAPI spec objects that we pull in from the .json files.
The merge is still shallow but it is deep "enough" now that it
does a merging of paths and schema components so that the
core-api and cmd-api-server openapi.json spec files can both be
used for request validation purposes.

The metadata lookups (used by the validation) were failing because
the shallow object merge was lossy, e.g. only one of the specs was
actually included by the validator.

Fixes #35

Signed-off-by: Peter Somogyvari <peter.metz@unarin.com>